Complete repository

+++ WebServer/src/main/kotlin/ru/hse/sd/webserver/ApiRequest.kt
@@ -1,13 +1,12 @@
 package ru.hse.sd.webserver
 
-import ru.hse.sd.hwproj.facade.Facade
-import ru.hse.sd.hwproj.model.Homework
-import ru.hse.sd.hwproj.model.RunnerTask
-import ru.hse.sd.hwproj.model.Submission
-import ru.hse.sd.hwproj.settings.WebServerApiSettings
 import java.net.URL
 import java.sql.Timestamp
 import kotlin.random.Random
+import ru.hse.sd.hwproj.facade.Facade
+import ru.hse.sd.hwproj.model.Homework
+import ru.hse.sd.hwproj.model.Submission
+import ru.hse.sd.hwproj.settings.WebServerApiSettings
 
 internal object ApiRequest : Facade(WebServerApiSettings) {
     // todo remove
@@ -24,26 +23,26 @@ internal object ApiRequest : Facade(WebServerApiSettings) {
 //    return request("/api/student/homework", HttpMethod.Post, homeworkContent)
         hw.add(
             Homework(
-                Random.nextLong(),
+                Random.nextInt(),
                 homeworkContent
             )
         )
         return hw.last()
     }
 
-    suspend fun getHomework(id: Long): Homework {
+    suspend fun getHomework(id: Int): Homework {
 //     TODO
 //    return request("/api/student/homework/${id}", HttpMethod.Get)
         return hw.single { it.id == id }
     }
 
-    suspend fun getSubmissions(id: Long): List<Submission> {
+    suspend fun getSubmissions(id: Int): List<Submission> {
 //     TODO
 //    return request(TODO, HttpMethod.Get)
         return List(Random.nextInt(7)) {
             Submission(
-                it.toLong(), Submission.Content(
-                    0L,
+                it.toInt(), Submission.Content(
+                    0,
                     Timestamp(Random.nextInt().toLong()),
                     URL("https://github.com/submission${Random.nextInt()}")
                 )
